+++ create-next-shadmag/example/dot-pattern-demo.tsx
@@ -1,0 +1,19 @@
+"use client";
+
+import { cn } from "@/lib/utils";
+import DotPattern from "@/components/magicui/dot-pattern";
+
+export default function DotPatternDemo() {
+  return (
+    <div className="relative flex h-[500px] w-full flex-col items-center justify-center overflow-hidden rounded-lg border border-neutral-200 bg-white md:shadow-xl dark:border-neutral-800 dark:bg-neutral-950">
+      <p className="z-10 whitespace-pre-wrap text-center text-5xl font-medium tracking-tighter text-black dark:text-white">
+        Dot Pattern
+      </p>
+      <DotPattern
+        className={cn(
+          "[mask-image:radial-gradient(300px_circle_at_center,white,transparent)]",
+        )}
+      />
+    </div>
+  );
+}