+++ create-next-shadmag/example/wavy-dot-pattern-demo.tsx
@@ -1,0 +1,18 @@
+// import WavyDotPattern from "@/components/magicui/wavy-dot-pattern";
+
+// export default function WavyDotPatternDemo() {
+//   return (
+//     <div className="relative flex h-full w-full max-w-[32rem] items-center justify-center overflow-hidden rounded-lg border border-neutral-200 bg-white p-20 md:shadow-xl  dark:border-neutral-800 dark:bg-neutral-950">
+//       <p className="z-10 whitespace-pre-wrap text-center text-4xl font-semibold tracking-tighter text-black dark:text-white">
+//         Wavy Dot Pattern
+//       </p>
+
+//       <WavyDotPattern
+//         gridWidth={30}
+//         gridHeight={15}
+//         dotWidth={8}
+//         dotHeight={8}
+//       />
+//     </div>
+//   );
+// }